Moved enumerators closer to files
This is more of a "style thing" for maintainability and not directly related to the logger, but since I was going to need to make a couple more enumerators, I thought I should just move it now instead of making it worse.
